+++ src/main/java/mod/azure/doom/entity/projectiles/EnergyCellEntity.java
@@ -242,7 +242,7 @@ public class EnergyCellEntity extends AbstractArrowEntity implements IAnimatable
 				((LivingEntity) entity1).setLastHurtMob(entity);
 			}
 		}
-		if (entity.hurt(damagesource, DoomConfig.SERVER.energycell_damage.get())) {
+		if (entity.hurt(damagesource, DoomConfig.SERVER.energycell_damage.get().floatValue())) {
 			if (entity instanceof LivingEntity) {
 				LivingEntity livingentity = (LivingEntity) entity;
 				if (!this.level.isClientSide && entity1 instanceof LivingEntity) {
