+++ src/main/java/mod/azure/doom/entity/projectiles/ShotgunShellEntity.java
@@ -249,7 +249,7 @@ public class ShotgunShellEntity extends AbstractArrowEntity implements IAnimatab
 				((LivingEntity) entity1).setLastHurtMob(entity);
 			}
 		}
-		if (entity.hurt(damagesource, DoomConfig.SERVER.shotgun_damage.get())) {
+		if (entity.hurt(damagesource, DoomConfig.SERVER.shotgun_damage.get().floatValue())) {
 			if (entity instanceof LivingEntity) {
 				LivingEntity livingentity = (LivingEntity) entity;
 				if (!this.level.isClientSide && entity1 instanceof LivingEntity) {
