+++ src/main/java/mod/azure/doom/entity/tierambient/TentacleEntity.java
@@ -180,7 +180,7 @@ public class TentacleEntity extends DemonEntity implements IAnimatable {
 			double d12 = (double) (MathHelper.sqrt(entity.distanceToSqr(vector3d)) / f2);
 			if (d12 <= 2.0D) {
 				if (entity instanceof LivingEntity) {
-					entity.hurt(DamageSource.indirectMagic(this, this.getTarget()), config.tentacle_melee_damage.get());
+					entity.hurt(DamageSource.indirectMagic(this, this.getTarget()), config.tentacle_melee_damage.get().floatValue());
 				}
 			}
 		}
