+++ src/main/java/mod/azure/doom/entity/tierfodder/MaykrDroneEntity.java
@@ -131,7 +131,7 @@ public class MaykrDroneEntity extends DemonEntity implements IAnimatable {
 		this.goalSelector.addGoal(4,
 				new RangedStrafeAttackGoal(this,
 						new MaykrDroneEntity.FireballAttack(this).setProjectileOriginOffset(0.8, 0.8, 0.8)
-								.setDamage(config.maykrdrone_ranged_damage.get()),
+								.setDamage(config.maykrdrone_ranged_damage.get().floatValue()),
 						1.0D, 50, 30, 15, 15F, 1).setMultiShot(2, 3));
 		this.targetSelector.addGoal(2, new NearestAttackableTargetGoal<>(this, PlayerEntity.class, true));
 		this.targetSelector.addGoal(3, new NearestAttackableTargetGoal<>(this, AbstractVillagerEntity.class, true));
