+++ src/main/java/mod/azure/doom/entity/tierheavy/BloodMaykrEntity.java
@@ -119,7 +119,7 @@ public class BloodMaykrEntity extends DemonEntity implements IAnimatable {
 		this.goalSelector
 				.addGoal(4,
 						new RangedStrafeAttackGoal(this, new BloodMaykrEntity.FireballAttack(this)
-								.setProjectileOriginOffset(0.8, 0.8, 0.8).setDamage(config.bloodmaykr_ranged_damage.get()), 1.0D, 50, 30, 15, 15F, 1)
+								.setProjectileOriginOffset(0.8, 0.8, 0.8).setDamage(config.bloodmaykr_ranged_damage.get().floatValue()), 1.0D, 50, 30, 15, 15F, 1)
 										.setMultiShot(2, 3));
 		this.goalSelector.addGoal(7, new BloodMaykrEntity.LookAroundGoal(this));
 		this.goalSelector.addGoal(5, new RandomFlyConvergeOnTargetGoal(this, 2, 15, 0.5));
