+++ src/main/java/mod/azure/doom/entity/tierheavy/MancubusEntity.java
@@ -180,7 +180,7 @@ public class MancubusEntity extends DemonEntity implements IAnimatable {
 				float f = (float) MathHelper.atan2(livingentity.getZ() - parentEntity.getZ(),
 						livingentity.getX() - parentEntity.getX());
 				BarenBlastEntity fireballentity = new BarenBlastEntity(world, this.parentEntity, d2, d3, d4,
-						config.mancubus_ranged_damage.get());
+						config.mancubus_ranged_damage.get().floatValue());
 				if (this.attackTimer == 15) {
 					if (parentEntity.distanceTo(livingentity) < 3.0D) {
 						for (int i = 0; i < 5; ++i) {
