+++ src/main/java/mod/azure/doom/entity/tiersuperheavy/Cyberdemon2016Entity.java
@@ -116,7 +116,7 @@ public class Cyberdemon2016Entity extends DemonEntity implements IAnimatable {
 		this.goalSelector.addGoal(5, new WaterAvoidingRandomWalkingGoal(this, 0.8D));
 		this.goalSelector.addGoal(4,
 				new RangedStaticAttackGoal(this, new Cyberdemon2016Entity.FireballAttack(this)
-						.setProjectileOriginOffset(0.8, 0.8, 0.8).setDamage(config.cyberdemon2016_ranged_damage.get()),
+						.setProjectileOriginOffset(0.8, 0.8, 0.8).setDamage(config.cyberdemon2016_ranged_damage.get().floatValue()),
 						60, 20, 30F, 1));
 		this.goalSelector.addGoal(4, new DemonAttackGoal(this, 1.0D, false, 2));
 		this.targetSelector.addGoal(2, new NearestAttackableTargetGoal<>(this, PlayerEntity.class, true));
