+++ src/main/java/mod/azure/doom/entity/tiersuperheavy/CyberdemonEntity.java
@@ -95,7 +95,7 @@ public class CyberdemonEntity extends DemonEntity implements IAnimatable {
 		this.goalSelector.addGoal(5, new WaterAvoidingRandomWalkingGoal(this, 0.8D));
 		this.goalSelector.addGoal(4,
 				new RangedStaticAttackGoal(this, new CyberdemonEntity.FireballAttack(this)
-						.setProjectileOriginOffset(0.8, 0.8, 0.8).setDamage(config.cyberdemon_ranged_damage.get()), 60,
+						.setProjectileOriginOffset(0.8, 0.8, 0.8).setDamage(config.cyberdemon_ranged_damage.get().floatValue()), 60,
 						20, 30F, 1));
 		this.targetSelector.addGoal(1, new HurtByTargetGoal(this).setAlertOthers());
 		this.targetSelector.addGoal(2, new NearestAttackableTargetGoal<>(this, PlayerEntity.class, true));
