+++ src/main/java/mod/azure/doom/entity/tiersuperheavy/FireBaronEntity.java
@@ -113,7 +113,7 @@ public class FireBaronEntity extends DemonEntity implements IAnimatable {
 		this.goalSelector.addGoal(5, new WaterAvoidingRandomWalkingGoal(this, 0.8D));
 		this.goalSelector.addGoal(4,
 				new RangedStaticAttackGoal(this, new FireBaronEntity.FireballAttack(this)
-						.setProjectileOriginOffset(0.8, 0.8, 0.8).setDamage(config.baron_ranged_damage.get()), 60, 20,
+						.setProjectileOriginOffset(0.8, 0.8, 0.8).setDamage(config.baron_ranged_damage.get().floatValue()), 60, 20,
 						30F, 2));
 		this.goalSelector.addGoal(4, new DemonAttackGoal(this, 1.20D, false, 1));
 		this.targetSelector.addGoal(2, new NearestAttackableTargetGoal<>(this, PlayerEntity.class, true));
